+++ packages/terra-form-select/tests/wdio/_frame-spec.js
@@ -1,10 +1,7 @@
-const viewports = Terra.viewports('tiny');
-
-describe('Frame', () => {
+Terra.describeViewports('Frame', ['tiny'], () => {
   describe('should have a themeable disabled icon', () => {
     before(() => browser.url('/#/raw/tests/terra-form-select/form-select/disabled'));
 
-    Terra.it.isAccessible({ viewports });
-    Terra.it.matchesScreenshot({ viewports });
+    Terra.it.validatesElement();
   });
 });