+++ packages/terra-hyperlink/tests/wdio/hyperlink-spec.js
@@ -1,16 +1,11 @@
-describe('Hyperlink', () => {
-  before(() => {
-    browser.setViewportSize(Terra.viewports('tiny')[0]);
-  });
-
+Terra.describeViewports('Hyperlink', ['tiny'], () => {
   describe('Default', () => {
     before(() => {
       browser.url('/#/raw/tests/terra-hyperlink/hyperlink/default-hyperlink');
       browser.moveToObject('#root', 0, 900); // move mouse to prevent hover styles
     });
 
-    Terra.it.isAccessible();
-    Terra.it.matchesScreenshot();
+    Terra.it.validatesElement();
   });
 
   describe('Disabled', () => {
@@ -21,70 +16,60 @@ describe('Hyperlink', () => {
       'color-contrast': { enabled: false },
     };
 
-    Terra.it.isAccessible({ rules });
-    Terra.it.matchesScreenshot();
+    Terra.it.validatesElement({ rules });
   });
 
   describe('Audio', () => {
     before(() => browser.url('/#/raw/tests/terra-hyperlink/hyperlink/audio-hyperlink'));
 
-    Terra.it.isAccessible();
-    Terra.it.matchesScreenshot();
+    Terra.it.validatesElement();
   });
 
   describe('Document', () => {
     before(() => browser.url('/#/raw/tests/terra-hyperlink/hyperlink/document-hyperlink'));
 
-    Terra.it.isAccessible();
-    Terra.it.matchesScreenshot();
+    Terra.it.validatesElement();
   });
 
   describe('External', () => {
     before(() => browser.url('/#/raw/tests/terra-hyperlink/hyperlink/external-hyperlink'));
 
-    Terra.it.isAccessible();
-    Terra.it.matchesScreenshot();
+    Terra.it.validatesElement();
   });
 
   describe('Image', () => {
     before(() => browser.url('/#/raw/tests/terra-hyperlink/hyperlink/image-hyperlink'));
 
-    Terra.it.isAccessible();
-    Terra.it.matchesScreenshot();
+    Terra.it.validatesElement();
   });
 
   describe('Video', () => {
     before(() => browser.url('/#/raw/tests/terra-hyperlink/hyperlink/video-hyperlink'));
 
-    Terra.it.isAccessible();
-    Terra.it.matchesScreenshot();
+    Terra.it.validatesElement();
   });
 
   describe('Icon Non Wrapping', () => {
     before(() => browser.url('/#/raw/tests/terra-hyperlink/hyperlink/icon-non-wrapping-hyperlink'));
 
-    Terra.it.isAccessible();
-    Terra.it.matchesScreenshot();
+    Terra.it.validatesElement();
   });
 
   describe('Scaled', () => {
     before(() => browser.url('/#/raw/tests/terra-hyperlink/hyperlink/scaled-hyperlink'));
 
-    Terra.it.isAccessible();
-    Terra.it.matchesScreenshot();
+    Terra.it.validatesElement();
   });
 
   describe('States', () => {
     before(() => browser.url('/#/raw/tests/terra-hyperlink/hyperlink/states-hyperlink'));
 
-    Terra.it.isAccessible();
-    Terra.it.matchesScreenshot();
+    Terra.it.validatesElement();
   });
 
   describe('Underline Hidden', () => {
     before(() => browser.url('/#/raw/tests/terra-hyperlink/hyperlink/underline-hidden-hyperlink'));
 
-    Terra.it.isAccessible();
-    Terra.it.matchesScreenshot();
+    Terra.it.validatesElement();
   });
 });