+++ packages/terra-action-header/CHANGELOG.md
@@ -3,6 +3,9 @@ ChangeLog
 
 Unreleased
 ----------
+
+2.18.0 - (July 9, 2019)
+------------------
 ### Changed
 * Updated to interface with react-intl's `injectIntl` removing `react's` legacy `contextTypes`.
 * Update tests to use `Terra.describeViewports` for setting viewports