+++ packages/terra-alert/CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 Unreleased
 ----------
+
+4.1.0 - (July 9, 2019)
+------------------
 ### Changed
 * Updated to interface with react-intl's `injectIntl` removing legacy `contextTypes`
 * Update tests to use `Terra.describeViewports` for setting viewports