+++ packages/terra-demographics-banner/CHANGELOG.md
@@ -5,6 +5,12 @@ Unreleased
 ----------
 ### Changed
 * Updated to interface with react-intl's `injectIntl` removing legacy `contextTypes`
+* Update tests to use `Terra.describeViewports` for setting viewports
+
+3.15.0 - (July 9, 2019)
+------------------
+### Changed
+* Minor dependency version bump
 
 3.14.0 - (June 28, 2019)
 ------------------