+++ packages/terra-form-checkbox/CHANGELOG.md
@@ -5,6 +5,12 @@ Unreleased
 ----------
 ### Changed
 * Updated to interface with react-intl's `injectIntl` removing legacy `contextTypes`
+* Update tests to use `Terra.describeViewports` for setting viewports and use `Terra.it.validatesElement`
+
+3.14.0 - (July 9, 2019)
+------------------
+### Changed
+* Minor dependency version bump
 
 3.13.0 - (June 28, 2019)
 ------------------