+++ packages/terra-overlay/CHANGELOG.md
@@ -5,6 +5,9 @@ Changelog
 
 Unreleased
 ----------
+
+3.18.0 - (July 9, 2019)
+------------------
 ### Fixed
 * Added `data-overlay-count` attribute to root document when overlay is opened
 * Added check for `data-modal-count` attribute on root document to track opening and closing of overlays and modals,