+++ packages/terra-search-field/CHANGELOG.md
@@ -6,6 +6,8 @@ Unreleased
 ### Changed
 * Updated to interface with react-intl's `injectIntl` removing legacy `contextTypes`
 
+3.17.0 - (July 9, 2019)
+------------------
 ### Fixed
 * Updated search field to pass event object to 'handleClear' from 'handleKeyDown'.